// src/lib/types/generated/Examinee.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntityId } from "./EntityId";

export type Examinee = { id: EntityId, nif: string, name: string, surenames: string, origin: string, court: number, academicCentreId: EntityId | null, subjectsIds: Array<EntityId>, };