// src/lib/types/generated/Subject.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntityId } from "./EntityId";
import type { SubjectKind } from "./SubjectKind";

export type Subject = { id: EntityId, name: string, kind: SubjectKind, examDate: string | null, examDuration: string | null, };