// src/lib/types/generated/Vigilant.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntityId } from "./EntityId";
import type { VigilantRole } from "./VigilantRole";

export type Vigilant = { id: EntityId, name: string, surenames: string, role: VigilantRole, specialtiesIds: Array<EntityId>, academicCentreId: EntityId | null, mainCourt: number, };