// src/lib/types/generated/Classroom.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntityId } from "./EntityId";

export type Classroom = { id: EntityId, code: string, locationCode: string, totalCapacity: number, examCapacity: number, priority: number, courtLocation: number | null, kind: string, notes: Array<string>, };